+++ Sources/Sentry/Public/SentryBaggage.h
@@ -41,13 +41,6 @@ NS_SWIFT_NAME(Baggage)
 @property (nullable, nonatomic, readonly) NSString *userId;
 
 /**
- * The value of a segment attribute in the user's data bag, if it exists.
- */
-#if !SDK_V9
-@property (nullable, nonatomic, readonly) NSString *userSegment;
-#endif
-
-/**
  * The random value used to determine if the trace is sampled.
  *
  * A float (`0.1234` notation) in the range of `[0, 1)` (including 0.0, excluding 1.0).
@@ -71,9 +64,6 @@ NS_SWIFT_NAME(Baggage)
                     releaseName:(nullable NSString *)releaseName
                     environment:(nullable NSString *)environment
                     transaction:(nullable NSString *)transaction
-#if !SDK_V9
-                    userSegment:(nullable NSString *)userSegment
-#endif
                      sampleRate:(nullable NSString *)sampleRate
                         sampled:(nullable NSString *)sampled
                        replayId:(nullable NSString *)replayId;
@@ -83,9 +73,6 @@ NS_SWIFT_NAME(Baggage)
                     releaseName:(nullable NSString *)releaseName
                     environment:(nullable NSString *)environment
                     transaction:(nullable NSString *)transaction
-#if !SDK_V9
-                    userSegment:(nullable NSString *)userSegment
-#endif
                      sampleRate:(nullable NSString *)sampleRate
                      sampleRand:(nullable NSString *)sampleRand
                         sampled:(nullable NSString *)sampled