+++ Sources/Sentry/SentrySDKInternal.m
@@ -704,9 +704,12 @@ static NSDate *_Nullable startTimestamp = nil;
             @"properties, so you can also just remove those lines from your configuration "
             @"altogether) before attempting to start a continuous profiling session. This behavior "
             @"relies on deprecated options and will change in a future version.");
+#    else
+    if (options == nil) {
+        SENTRY_LOG_WARN(@"Cannot start profiling when options are nil.");
+#    endif
         return;
     }
-#    endif // !SDK_V9
 
     if (options.profiling != nil) {
         if (options.profiling.lifecycle == SentryProfileLifecycleTrace) {
@@ -772,9 +775,12 @@ static NSDate *_Nullable startTimestamp = nil;
             @"properties, so you can also just remove those lines from your configuration "
             @"altogether) before attempting to stop a continuous profiling session. This behavior "
             @"relies on deprecated options and will change in a future version.");
+#    else
+    if (options == nil) {
+        SENTRY_LOG_WARN(@"Cannot stop profiling when options are nil.");
+#    endif
         return;
     }
-#    endif // !SDK_V9
 
     if (options.profiling != nil && options.profiling.lifecycle == SentryProfileLifecycleTrace) {
         SENTRY_LOG_WARN(