+++ Tests/SentryTests/Protocol/SentryDebugMetaEquality.swift
@@ -3,10 +3,8 @@ import Foundation
 extension DebugMeta {
     open override func isEqual(_ object: Any?) -> Bool {
         if  let other = object as? DebugMeta {
-            return  uuid == other.uuid &&
-            debugID == other.debugID &&
+            return debugID == other.debugID &&
                 type == other.type &&
-                name == other.name &&
                 codeFile == other.codeFile &&
                 imageSize == other.imageSize &&
                 imageAddress == other.imageAddress &&