+++ Tests/SentryTests/Protocol/TestData.swift
@@ -85,10 +85,8 @@ class TestData {
         debugMeta.imageAddress = "0x0000000105705000"
         debugMeta.imageSize = 352_256
         debugMeta.imageVmAddress = "0x00007fff51af0000"
-        debugMeta.name = "/tmp/scratch/dyld_sim"
         debugMeta.codeFile = "/tmp/scratch/dyld_sim"
         debugMeta.type = "macho"
-        debugMeta.uuid = "84BAEBDA-AD1A-33F4-B35D-8A45F5DAF322"
         debugMeta.debugID = "84BAEBDA-AD1A-33F4-B35D-8A45F5DAF321"
         
         return debugMeta
@@ -244,7 +242,7 @@ class TestData {
 
     static var debugImage: DebugMeta {
         let image = DebugMeta()
-        image.name = "sentrytest"
+        image.codeFile = "sentrytest"
         image.imageAddress = "0x0000000105705000"
         image.imageVmAddress = "0x0000000105705000"
         return image